new input, proxy pointer system, and more
Mesh work has started (barely)
New pointer proxy system (needs extensive testing)
HID mouse support now for the camera
Transforms now work via proxy pointers

+++ tool/src/Input.cpp
@@ -113,7 +113,7 @@ namespace Input
 
 	void OnMouseMove(int x, int y)
 	{
-		mouseDelta.x = x;
-		mouseDelta.y = y;
+		mouseDelta.x = (float)x;
+		mouseDelta.y = (float)y;
 	}
 }
